requirements.txt
requests>=2.28
# tenacity could be used for more advanced retry/backoff, but current implementation uses a simple custom loop
python-stdnum>=1.16
# python-stdnum can be used for IBAN validation: python-stdnum>=1.16
pytest>=7.0
responses>=0.23.0

